requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.1.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==1.7.3
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.3.2
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2